+++ src/Detach/Collisions/Geometry2D.cs
@@ -40,7 +40,7 @@ public static class Geometry2D
 
 	public static bool PointInOrientedRectangle(Vector2 point, OrientedRectangle orientedRectangle)
 	{
-		Vector2 rotVector = point - orientedRectangle.Position;
+		Vector2 rotVector = point - orientedRectangle.Center;
 		float theta = -orientedRectangle.RotationInRadians;
 		Matrix2 zRotation = new(
 			MathF.Cos(theta), MathF.Sin(theta),
@@ -136,8 +136,8 @@ public static class Geometry2D
 			-MathF.Sin(theta), MathF.Cos(theta));
 
 		LineSegment2D localLine = new(
-			Matrices.Multiply(line.Start - orientedRectangle.Position, zRotation) + orientedRectangle.HalfExtents,
-			Matrices.Multiply(line.End - orientedRectangle.Position, zRotation) + orientedRectangle.HalfExtents);
+			Matrices.Multiply(line.Start - orientedRectangle.Center, zRotation) + orientedRectangle.HalfExtents,
+			Matrices.Multiply(line.End - orientedRectangle.Center, zRotation) + orientedRectangle.HalfExtents);
 		Rectangle localRectangle = Rectangle.FromTopLeft(Vector2.Zero, orientedRectangle.HalfExtents * 2);
 		return LineRectangle(localLine, localRectangle);
 	}
@@ -188,7 +188,7 @@ public static class Geometry2D
 			-MathF.Sin(theta), MathF.Cos(theta));
 
 		Circle localCircle = new(
-			Matrices.Multiply(circle.Position - orientedRectangle.Position, zRotation) + orientedRectangle.HalfExtents,
+			Matrices.Multiply(circle.Position - orientedRectangle.Center, zRotation) + orientedRectangle.HalfExtents,
 			circle.Radius);
 		Rectangle localRectangle = Rectangle.FromTopLeft(Vector2.Zero, orientedRectangle.HalfExtents * 2);
 		return CircleRectangle(localCircle, localRectangle);
@@ -273,8 +273,8 @@ public static class Geometry2D
 	public static bool OrientedRectangleOrientedRectangleSat(OrientedRectangle orientedRectangle1, OrientedRectangle orientedRectangle2)
 	{
 		Rectangle local1 = Rectangle.FromTopLeft(Vector2.Zero, orientedRectangle1.HalfExtents * 2);
-		Vector2 rotVector = orientedRectangle2.Position - orientedRectangle1.Position;
-		OrientedRectangle local2 = new(orientedRectangle2.Position, orientedRectangle2.HalfExtents, orientedRectangle2.RotationInRadians);
+		Vector2 rotVector = orientedRectangle2.Center - orientedRectangle1.Center;
+		OrientedRectangle local2 = new(orientedRectangle2.Center, orientedRectangle2.HalfExtents, orientedRectangle2.RotationInRadians);
 		local2.RotationInRadians -= orientedRectangle1.RotationInRadians;
 
 		float theta = -orientedRectangle1.RotationInRadians;
@@ -283,14 +283,14 @@ public static class Geometry2D
 			-MathF.Sin(theta), MathF.Cos(theta));
 
 		rotVector = Matrices.Multiply(rotVector, zRotation);
-		local2.Position = rotVector + orientedRectangle1.HalfExtents;
+		local2.Center = rotVector + orientedRectangle1.HalfExtents;
 
 		return RectangleOrientedRectangleSat(local1, local2);
 	}
 
 	public static bool OrientedRectangleTriangle(OrientedRectangle orientedRectangle, Triangle2D triangle)
 	{
-		if (PointInTriangle(orientedRectangle.Position, triangle))
+		if (PointInTriangle(orientedRectangle.Center, triangle))
 			return true;
 
 		LineSegment2D ab = new(triangle.A, triangle.B);